Index
Create an HTML page called "index.html" and an external CSS stylesheet called "styles.css". Set the body to use a different font than the default and a different background color than the default. Create a header, "Hello World" with an ID of "main-title" and style that ID to be centered. Create two paragraphs, each with a class of "content" and apply a bottom margin of 12px to that class.